feat(records): the VAT number carries its own verdict
The check lived on the Profile tab, inside a collapsed section, two tabs from
the rail where the number itself is. Somebody looking straight at a company's
VAT ID never met the thing that checks it — which was reported as exactly that:
"I don't see anything here", from a reader with the field on screen.

The verdict is now a mark beside the number. A shield in the status colours
says valid, not valid, or not checked yet, readable without opening anything,
because whether a stated number holds up is the fact somebody wants at a glance
and the consultation is what they want occasionally. A number a copied imprint
took from another company is the finding this exists for, and it should not
cost a click.

The mark opens to what the old card held: the register's answer, the number
consulted, the name the register holds — a number validating to a company
nobody recognises is the finding — the date, the receipt, and the button that
asks again.

The Profile section is gone rather than kept alongside. Two surfaces for one
fact is how they drift, and the one that stays is the one where both halves are
in a single glance. Its two headings went with it: the catalogue gate refuses a
key translated three times and rendered nowhere, which is how they were found.

One thing the card could not say, because it never sat beside the number: the
row is editable, so a receipt can end up next to a number nobody consulted. The
mark checks, and says the number moved.
